requirements.txt
fastapi[standard]==0.115.13
psycopg=3.2.9